Crate decaf377_fmd

An implementation of Fuzzy Message Detection.

Structs

Clue

A clue that allows probabilistic message detection.

ClueKey

Bytes representing a clue key corresponding to some DetectionKey.

DetectionKey

Used to examine Clues and determine whether they were possibly sent to the detection key’s ClueKey.

ExpandedClueKey

An expanded and validated clue key that can be used to create Clues intended for the corresponding DetectionKey.

Precision

Represents the precision governing the false positive rate of detection.

Enums

Error

An error in message detection.